Add rubocop and bundler-audit

+++ db/migrate/20180529125011_replace_id_to_code.rb
@@ -32,9 +32,7 @@ class ReplaceIdToCode < ActiveRecord::Migration[4.2]
       Currency.update_all('id = code')
       remove_column :currencies, :code
 
-      if index_exists?(:currencies, %i[enabled code])
-        remove_index :currencies, column: %i[enabled code]
-      end
+      remove_index :currencies, column: %i[enabled code] if index_exists?(:currencies, %i[enabled code])
 
       add_index :currencies, [:enabled]
     end
